+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useState } from 'react';
 import styled from 'styled-components';
 
-import useUrlState from './useUrlStateNew';
+import useUrlState from './useUrlState';
 
 import {
   ActionBar,
